Initial commit, Dockerfile and Makefile
The Dockerfile extends the official Postgres Alpine images with two init
scripts to create the databases and users needed and set the DataStore
DB permissions.

It also changes the location of the Postgres data file (`PGDATA`) so
they are created inside the image and not on a mounted volume. This
allows us to:
1. Build this base image
2. Start a container using this base image, which will run the init
   scripts
3. Commit changes to a new image that includes the postgres changes,
   which is the one that we will publish

All a bit hacky, but we'll make clear that these images should only be
used for testing.

The Makefile automates the build process and allows to build different
versions.